Fix PR #6 CI: remove duplicate GatewayConnectionStore
